change gearhash tests to wasm
