+++ packages/gearhash-wasm/vitest.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+	test: {
+		globals: true,
+		environment: "node",
+	},
+	resolve: {
+		alias: {
+			"@huggingface/splitmix64-wasm": "./node_modules/@huggingface/splitmix64-wasm/build/release.js",
+		},
+	},
+});
